refactor: rename storage_manager csrc dir and namespace to lmcache_native Signed-off-by: baoloongmao <307499405@qq.com>
Rename csrc/storage_manager to csrc/lmcache_native and the
lmcache::storage_manager C++ namespace to lmcache::lmcache_native
to match the existing pybind module name (lmache_native). Update
build config (setup_extensions/common_cpp.py) and comments.
