+++ csrc/lmcache_native/utils.cpp
