+++ csrc/lmcache_native/utils.h
